backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-huggingface>=0.1.0

chromadb>=0.4.0
sentence-transformers>=2.2.0

pypdf>=3.17.0
python-markdown>=3.5.0

SQLAlchemy>=2.0.0
alembic>=1.13.0

structlog>=24.1.0
PyYAML>=6.0

httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
